# sstables: enable read ahead for the partition index reader