@denesb 2021-07-21
> Currently, when advancing one of `index_reader`'s bounds, we're creating a new `index_consume_entry_context` with a new underlying file `input_stream` for each new page.
> 
> For either bound, the streams can be reused, because the indexes of pages that we are reading are never decreasing.

This might be the case today but we have plans to change this to support [reverse reads](https://github.com/scylladb/scylla/issues/1413). What I want to say is that although it is fine to build on this assumption to speed up the common fast path of forward reads, we should not make reading in a zig-zag pattern impossible. I am saying this without having reviewed the code, so this might be a moot point, but wanted to state it nevertheless.