@avikivity 2021-07-21
I can't say I can follow the intricacies of the index reader, but overall looks good, especially the results.

Q: how does this affect single small partition read? Will it cause larger I/Os on the index?

I think you can obtain an answer with tracing, it will show file offsets and sizes on which we initate I/O.