@wmitros 2021-07-21
> > Currently, when advancing one of `index_reader`'s bounds, we're creating a new `index_consume_entry_context` with a new underlying file `input_stream` for each new page.
> > For either bound, the streams can be reused, because the indexes of pages that we are reading are never decreasing.
> 
> This might be the case today but we have plans to change this to support [reverse reads](https://github.com/scylladb/scylla/issues/1413). What I want to say is that although it is fine to build on this assumption to speed up the common fast path of forward reads, we should not make reading in a zig-zag pattern impossible. I am saying this without having reviewed the code, so this might be a moot point, but wanted to state it nevertheless.

You're right that this patch is mostly about speeding up the common fast path, but I'm also going to work on the reverse reads, so I'll add my 2 cents about the interaction:
As far as I know, we are going to focus on optimizing the reversed reads from a single partition, and the streams that this patch combines are used for reading the partition data, not the rows in a single partition (for that we have promoted index, which has a separate stream).
Besides, I haven't read much into the reasons why, but the index reader already assumes that we're reading pages in a non decreasing order as can be seen here: https://github.com/scylladb/scylla/blob/7261c2c93e633a5df2772cc1313da59712ba74d4/sstables/index_reader.hh#L485-L488 so it is not a new assumption in this patch