@avikivity 2021-07-21
> As far as I know, we are going to focus on optimizing the reversed reads from a single partition, and the streams that this patch combines are used for reading the partition data, not the rows in a single partition (for that we have promoted index, which has a 

Reversed reads are only defined within a single partition. This is because multi-partition reads return partitions in token order, which is arbitrary as far as the application is concerned, whereas clustering key order is meaningful.