@wmitros 2021-07-21
> I can't say I can follow the intricacies of the index reader, but overall looks good, especially the results.
> 
> Q: how does this affect single small partition read? Will it cause larger I/Os on the index?
> 
> I think you can obtain an answer with tracing, it will show file offsets and sizes on which we initate I/O.

I'll try to test it soon, but before that I think I should add that there is one improvement to this patch that I've attempted but didn't seem worth the extra code churn, which would help with decreasing the amount of I/Os on the index:
In this patch, in order to enable advancing the index read context, the underlying file input stream is set to cover the entire index file. The improvement is that we could limit it when we know that the associated sstable reader is not forwarding - in that case we would only need to set the limits of the input stream to the partition range of the sstable reader