@avikivity 2021-07-21
> > > As far as I know, we are going to focus on optimizing the reversed reads from a single partition, and the streams that this patch combines are used for reading the partition data, not the rows in a single partition (for that we have promoted index, which has a
> > 
> > 
> > Reversed reads are only defined within a single partition. This is because multi-partition reads return partitions in token order, which is arbitrary as far as the application is concerned, whereas clustering key order is meaningful.
> 
> Right, though we do allow reads with reversed order on clustering key for multiple partitions (WHERE pk IN ...), with paging disabled, but I assume after we implement reversed reads for single partition, we can perform the same process for multiple partitions

I imagine it returns the partitions in the same order as unreversed reads. So even if we implement it for multiple partitions (for which there is no reason to, at least yet), we'll probably do the same.

I can see some use for 

```sql
SELECT *
FROM tab
ORDER BY score DESC
LIMIT 10
```

to give the best 10 scores per partition. Nobody cares about the partition order, but the clustering key order (assuming score is a clustering key here, which is dubious) is important.